+++ src/modules/Toast/Toast.stories.tsx
@@ -8,9 +8,9 @@ const ToastStory = () => {
   const { addToast } = useToasts();
   return (
     <>
-      <Button onClick={() => addToast.success('success')} text="show success toast" />
-      <Button onClick={() => addToast.info('info')} text="show info toast" />
-      <Button onClick={() => addToast.error('error')} text="show error toast" />
+      <Button onClick={() => addToast.success('success')}>show success toast</Button>
+      <Button onClick={() => addToast.info('info')}>show info toast</Button>
+      <Button onClick={() => addToast.error('error')}>show error toast</Button>
     </>
   );
 };