+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Unreleased
 - Added MetaTable module
 
+### Button
+- removed `value` property
+- `error` property renamed to `danger`
+- removed `bordered` property
+- removed `id` property
+- content of a button passed as a children instead of `text` property
+
+### LinkButton
+- renamed to `Link`
+
+
 # 0.2.0
 - Removed Chevron, Label, Typography, Tags